os/memstore: Cleanup and refactor Object Classes
